refactor (Procedure): Port/Adapter 패턴 적용
