+++ src/main/java/com/sopt/cherrish/domain/procedure/domain/port/ProcedureSearchPort.java
@@ -1,0 +1,10 @@
+package com.sopt.cherrish.domain.procedure.domain.port;
+
+import java.util.List;
+
+public interface ProcedureSearchPort {
+
+	boolean isAvailable();
+
+	List<Long> searchByKeyword(String keyword);
+}
